refactor: 필요없는 state 제거 및 UI 수정
