+++ front-end/legeno-around-here/src/components/pages/myProfileEdit/PhotoEditSection.js
@@ -1,58 +1,30 @@
-import React, { useState } from 'react'
-import {MAIN_COLOR} from '../../../constants/Color';
-import makeStyles from '@material-ui/core/styles/makeStyles';
-import IconButton from '@material-ui/core/IconButton'
-import { saveProfilePhoto } from '../../api/API'
-import Typography from '@material-ui/core/Typography'
+import React from 'react';
+import IconButton from '@material-ui/core/IconButton';
+import { saveProfilePhoto } from '../../api/API';
+import Typography from '@material-ui/core/Typography';
+import useStyle from './PhotoEditSectionStyle';
 
-const useStyle = makeStyles({
-  photo: (props) => ({
-    width: '100px',
-    height: '100px',
-    backgroundRepeat: 'no-repeat',
-    backgroundSize: 'cover',
-    backgroundPosition: 'center',
-    borderRadius: '300px',
-    backgroundImage: `url('${props.photoUrl}')`,
-    border: '1px solid' + MAIN_COLOR,
-  }),
-  photoEditSection: {
-    width: '90%',
-    display: 'flex',
-    alignItems: 'center',
-    margin: '20px auto',
-  },
-  uploadPhoto: {
-    opacity: 0,
-    position: 'absolute',
-    zIndex: -1,
-    pointerEvents: 'none',
-  },
-  photoEditButton: {
-    color: 'black',
-  }
-});
-
-const PhotoEditSection = ({ originalPhotoUrl, setProfilePhoto, accessToken }) => {
-  const [photoUrl, setPhotoUrl] = useState(originalPhotoUrl);
-
+const PhotoEditSection = ({
+  originalPhotoUrl,
+  setProfilePhoto,
+  accessToken,
+}) => {
   const props = {
-    photoUrl: photoUrl,
+    photoUrl: originalPhotoUrl,
   };
   const classes = useStyle(props);
 
   const onImagesChanged = (e) => {
     const formData = new FormData();
     formData.append('image', e.target.files[0]);
-    saveProfilePhoto(formData, accessToken)
-      .then(responseData => {
-        setPhotoUrl(responseData.url);
+    saveProfilePhoto(formData, accessToken).then((responseData) => {
+      if (responseData) {
         setProfilePhoto({
           id: responseData.id,
-          url: responseData.url
+          url: responseData.url,
         });
-        alert(responseData.url + ", id: " + responseData.id);
-      });
+      }
+    });
   };
 
   return (
